Convert Services to typescript
